@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Net;
 using System.Text.Json;
-using Challonge.Api;
+using Challonge.JsonConverters;
 
 namespace Challonge.Exceptions
 {
@@ -21,7 +21,7 @@ namespace Challonge.Exceptions
 				HttpStatusCode.Unauthorized => "Unauthorized - Invalid credentials or insufficient permissions.",
 				HttpStatusCode.NotAcceptable => "Invalid response format specified. This is most likely an internal Challonge-DotNet error, please report an issue in the Github repository.",
 				HttpStatusCode.InternalServerError => "An unspecified Challonge server error occurred.",
-				HttpStatusCode.UnprocessableEntity or HttpStatusCode.NotFound => JsonSerializer.Deserialize<ErrorResponse>(responseText, ChallongeClient.JsonSerializerOptions)!.Message,
+				HttpStatusCode.UnprocessableEntity or HttpStatusCode.NotFound => JsonSerializer.Deserialize<ErrorResponse>(responseText, ChallongeJsonContext.Default.ErrorResponse)!.Message,
 				_ => responseText
 			})
 		{
